+++ params/types/genesisT/gen_genesis.go
@@ -29,7 +29,7 @@ import (
 	"github.com/ethereum/go-ethereum/params/confp/generic"
 	common0 "github.com/ethereum/go-ethereum/params/types/ctypes"
 	"github.com/ethereum/go-ethereum/params/types/goethereum"
-	"github.com/ethereum/go-ethereum/params/types/chippr-geth"
+	"github.com/ethereum/go-ethereum/params/types/chipprgeth"
 	"github.com/ethereum/go-ethereum/params/types/multigeth"
 )
